+++ src/Command/Data/Cli.php
@@ -3,13 +3,13 @@
 namespace ChessServer\Command\Data;
 
 use ChessServer\Command\AbstractCli;
-use ChessServer\Command\Data\Async\AnnotationsGameCommand;
-use ChessServer\Command\Data\Async\AutocompleteBlackCommand;
-use ChessServer\Command\Data\Async\AutocompleteEventCommand;
-use ChessServer\Command\Data\Async\AutocompleteWhiteCommand;
-use ChessServer\Command\Data\Async\OpeningCommand;
-use ChessServer\Command\Data\Async\RankingCommand;
-use ChessServer\Command\Data\Async\SearchCommand;
+use ChessServer\Command\Data\Sync\AnnotationsGameCommand;
+use ChessServer\Command\Data\Sync\AutocompleteBlackCommand;
+use ChessServer\Command\Data\Sync\AutocompleteEventCommand;
+use ChessServer\Command\Data\Sync\AutocompleteWhiteCommand;
+use ChessServer\Command\Data\Sync\OpeningCommand;
+use ChessServer\Command\Data\Sync\RankingCommand;
+use ChessServer\Command\Data\Sync\SearchCommand;
 use Spatie\Async\Pool;
 
 class Cli extends AbstractCli
