+++ src/Command/Data/Sync/AutocompleteWhiteCommand.php
@@ -1,26 +1,31 @@
 <?php
 
-namespace ChessServer\Command\Data\Async;
+namespace ChessServer\Command\Data\Sync;
 
 use ChessServer\Command\AbstractAsyncCommand;
 use ChessServer\Socket\AbstractSocket;
 
-class RankingCommand extends AbstractAsyncCommand
+class AutocompleteWhiteCommand extends AbstractAsyncCommand
 {
     public function __construct()
     {
-        $this->name = '/ranking';
-        $this->description = 'Top players by ELO.';
+        $this->name = '/autocomplete_white';
+        $this->description = 'Autocomplete data for chess players.';
+        $this->params = [
+            'params' => '<string>',
+        ];
     }
 
     public function validate(array $argv)
     {
-        return count($argv) - 1 === 0;
+        return count($argv) - 1 === count($this->params);
     }
 
     public function run(AbstractSocket $socket, array $argv, int $id)
     {
-        $this->pool->add(new RankingTask())
+        $params = json_decode(stripslashes($argv[1]), true);
+
+        $this->pool->add(new AutocompleteWhiteTask($params))
             ->then(function ($result) use ($socket, $id) {
                 return $socket->getClientStorage()->send([$id], [
                     $this->name => $result,
