+++ src/Command/Data/Sync/SearchCommand.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace ChessServer\Command\Data\Async;
+namespace ChessServer\Command\Data\Sync;
 
 use ChessServer\Command\AbstractAsyncCommand;
 use ChessServer\Socket\AbstractSocket;
 
-class OpeningCommand extends AbstractAsyncCommand
+class SearchCommand extends AbstractAsyncCommand
 {
     public function __construct()
     {
-        $this->name = '/opening';
-        $this->description = 'Opening results.';
+        $this->name = '/search';
+        $this->description = 'Finds up to 25 games matching the criteria.';
         $this->params = [
             'params' => '<string>',
         ];
@@ -25,7 +25,7 @@ class OpeningCommand extends AbstractAsyncCommand
     {
         $params = json_decode(stripslashes($argv[1]), true);
 
-        $this->pool->add(new OpeningTask($params), 128000)
+        $this->pool->add(new SearchTask($params), 128000)
             ->then(function ($result) use ($socket, $id) {
                 return $socket->getClientStorage()->send([$id], [
                     $this->name => $result,
